Add files via upload
Completed projects on Hugging Face, and I would like to thank Simplilearn for empowering me to move towards positive heights or growth, and it's my pleasure to continue with Simplilearn. And Simplilearn is a GOD to me, and I was happy to continue my code journey again after so many years and happy coding. trying for work-life balance and to move towards positive growth.OK, thank you, Simplilearn and thank you. Have a great day!!